requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.52.2
plotly==6.5.0
click==8.3.1
pydantic==2.10.6
sqlalchemy==2.0.23
faiss-cpu==1.13.1
lancedb==0.26.0
duckdb==1.4.3
rich==14.2.0
pandas==2.3.3
matplotlib==3.10.8
scikit-learn==1.7.2
pyyaml==6.0.2
pytest==7.4.3
python-dotenv==1.0.1
requests==2.32.5



